# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This service manages all backend interactions for the feed the realm clients.\n\n⚠️ NOTE: All 2xx responses are wrapped in a 'Data envelope'.\nExample: { \"data\": { \"id\": \"123\", \"name\": \"Alice\" } }",
        "title": "FeedTheRealm Core Service API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/exports/zip": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the path for a given app, version, and OS.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exports-service"
                ],
                "summary": "Get export zip path",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App name",
                        "name": "app",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version (vX.Y.Z)",
                        "name": "version",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS name",
                        "name": "os",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ExportZipPathResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Uploads a zip file for a given app, version, and OS.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exports-service"
                ],
                "summary": "Upload export zip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App name",
                        "name": "app",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version (vX.Y.Z)",
                        "name": "version",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OS name",
                        "name": "os",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Zip file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.ExportZipResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dtos.ExportZipPathResponse": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
        "dtos.ExportZipResponse": {
            "type": "object",
            "properties": {
                "app_name": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
    }
}